Seperate IgnoredPropertyAttributes into its own file
